refacto: move validation schemas from constants to db
